Refactor data load, fix pagination state, and improve select value binding
This commit updates the `tailwind-view.component.ts` to use an RXJS timer instead of a `setTimeout` for loading data. This change provides a more consistent and reliable way to handle asynchronous data loading.

Additionally, `angularx-datatable-tailwind.component.ts` has been updated to ensure that the current page in pagination doesn't reset to 1 each time, but instead retains its state even when the component re-renders. This was important to maintain the user's place in the paginated list.

Lastly, the `[value]` binding of the select option in the `angularx-datatable-tailwind.component.html` file was replaced with `[ngValue]